requirements.txt
absl-py==0.6.1
aniso8601==4.0.1
astor==0.7.1
Click==7.0
Flask==1.0.2
Flask-Redis==0.3.0
gast==0.2.0
grpcio==1.17.1
gunicorn==19.9.0
h5py==2.8.0
itsdangerous==1.1.0
Jinja2==2.10
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
Markdown==3.0.1
MarkupSafe==1.1.0
numpy==1.15.4
Pillow==5.3.0
protobuf==3.6.1
pytz==2018.7
PyYAML==3.13
redis==3.0.1
scipy==1.1.0
six==1.12.0
tensorboard==1.12.1
tensorflow==1.12.0
termcolor==1.1.0
Werkzeug==0.14.1
